Extract AutoYaST LVM planning to AutoinstVgPlanner
